requirements.txt
beautifulsoup4==4.10.0
lxml==4.6.3
pydantic==1.9.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot==13.7
redis==4.3.3
requests==2.26.0
schedule==1.1.0
